+++ nemo_nowcast/message_broker.py
@@ -21,8 +21,10 @@ import logging
 import logging.config
 import os
 import signal
+import time
 
 import zmq
+import zmq.log.handlers
 
 from nemo_nowcast import (
     CommandLineInterface,
@@ -45,7 +47,7 @@ def main():
       to launch the message broker
     * Reading and parsing the configuration file given on the command-line
     * Configuring the logging system as specified in the configuration file
-    * Logging the message broker's PID, and the file path/name that was used to
+    * Log the message broker's PID, and the file path/name that was used to
       configure it.
 
     The set-up is repeated if the message broker process receives a HUP signal
@@ -63,19 +65,59 @@ def main():
     parsed_args = cli.parser.parse_args()
     config = Config()
     config.load(parsed_args.config_file)
-    # Replace logging RotatingFileHandlers with WatchedFileHandlers so that we
-    # notice when log files are rotated and switch to writing to the new ones
-    logging_handlers = config['logging']['handlers']
-    rotating_handler = 'logging.handlers.RotatingFileHandler'
-    watched_handler = 'logging.handlers.WatchedFileHandler'
-    for handler in logging_handlers:
-        if logging_handlers[handler]['class'] == rotating_handler:
-            logging_handlers[handler]['class'] = watched_handler
-            del logging_handlers[handler]['backupCount']
-    logging.config.dictConfig(config['logging'])
+    msg = _configure_logging(config)
     logger.info('running in process {}'.format(os.getpid()))
     logger.info('read config from {.file}'.format(config))
+    logger.info(msg)
     run(config)
+
+
+def _configure_logging(config):
+    """Configure the message broker's logging system interface.
+
+    :param config: Nowcast system configuration.
+    :type config: :py:class:`nemo_nowcast.config.Config`
+    """
+    if 'publisher' in config['logging']:
+        # Publish log messages to distributed logging aggregator
+        logging_config = config['logging']['publisher']
+        logging_config['handlers']['zmq_pub']['context'] = context
+        host = config['zmq']['host']
+        port = config['zmq']['ports']['logging'][NAME]
+        addr = 'tcp://*:{port}'.format(port=port)
+        logging_config['handlers']['zmq_pub']['interface_or_socket'] = addr
+        logging.config.dictConfig(logging_config)
+        for handler in logger.root.handlers:
+            if isinstance(handler, zmq.log.handlers.PUBHandler):
+                handler.root_topic = NAME
+                handler.formatters = {
+                    logging.DEBUG: logging.Formatter("%(message)s\n"),
+                    logging.INFO: logging.Formatter("%(message)s\n"),
+                    logging.WARNING: logging.Formatter("%(message)s\n"),
+                    logging.ERROR: logging.Formatter("%(message)s\n"),
+                    logging.CRITICAL: logging.Formatter("%(message)s\n"),
+                }
+        # Not sure why, but we need a brief pause before we start logging
+        # messages
+        time.sleep(0.25)
+        msg = 'publishing logging messages to {addr}'.format(addr=addr)
+    else:
+        # Write log messages to local file system
+        #
+        # Replace logging RotatingFileHandlers with WatchedFileHandlers so
+        # that we notice when log files are rotated and switch to writing to
+        # the new ones
+        logging_config = config['logging']
+        logging_handlers = logging_config['handlers']
+        rotating_handler = 'logging.handlers.RotatingFileHandler'
+        watched_handler = 'logging.handlers.WatchedFileHandler'
+        for handler in logging_handlers:
+            if logging_handlers[handler]['class'] == rotating_handler:
+                logging_handlers[handler]['class'] = watched_handler
+                del logging_handlers[handler]['backupCount']
+        logging.config.dictConfig(logging_config)
+        msg = 'writing logging messages to local file system'
+    return msg
 
 
 def run(config):
@@ -85,6 +127,9 @@ def run(config):
       communication with the manager and worker processes.
     * Install signal handlers for hangup, interrupt, and kill signals.
     * Launch the brokers message queuing process.
+
+    :param config: Nowcast system configuration.
+    :type config: :py:class:`nemo_nowcast.config.Config`
     """
     workers_socket, manager_socket = _bind_zmq_sockets(config)
     _install_signal_handlers(workers_socket, manager_socket)
@@ -105,6 +150,9 @@ def run(config):
 
 def _bind_zmq_sockets(config):
     """Create 0mq sockets and bind them to ports.
+
+    :param config: Nowcast system configuration.
+    :type config: :py:class:`nemo_nowcast.config.Config`
     """
     workers_socket = context.socket(zmq.ROUTER)
     manager_socket = context.socket(zmq.DEALER)