+++ tests/test_message_broker.py
@@ -27,14 +27,17 @@ import zmq
 from nemo_nowcast import message_broker
 
 
-@patch('nemo_nowcast.message_broker.logging')
-@patch('nemo_nowcast.message_broker.Config')
 @patch('nemo_nowcast.message_broker.CommandLineInterface')
+@patch('nemo_nowcast.message_broker.Config')
+@patch('nemo_nowcast.message_broker._configure_logging')
+@patch('nemo_nowcast.message_broker.logging')
 @patch('nemo_nowcast.message_broker.run')
 class TestMain:
     """Unit tests for message_broker.main function.
     """
-    def test_commandline_interface(self, m_run, m_cli, m_config, m_logging):
+    def test_commandline_interface(
+        self, m_run, m_logging, m_config_logging, m_config, m_cli
+    ):
         message_broker.main()
         args, kwargs = m_cli.call_args_list[0]
         assert args[0] == 'message_broker'
@@ -42,47 +45,113 @@ class TestMain:
         assert 'description' in kwargs
         m_cli.build_parser.asser_called_once_with()
 
-    def test_cli_parser(self, m_run, m_cli, m_config, m_logging):
+    def test_cli_parser(
+        self, m_run, m_logging, m_config_logging, m_config, m_cli
+    ):
         message_broker.main()
         m_cli().parser.parse_args.assert_called_once_with()
 
-    def test_config_load(self, m_run, m_cli, m_config, m_logging):
-        m_cli().parser.parse_args.return_value = Mock(
-            config_file='nowcast.yaml')
-        message_broker.main()
-        m_config().load.assert_called_once_with('nowcast.yaml')
-
-    def test_change_rotating_logger_handler_to_watched(
-        self, m_run, m_cli, m_config, m_logging
+    def test_config_load(
+        self, m_run, m_logging, m_config_logging, m_config, m_cli
     ):
         m_cli().parser.parse_args.return_value = Mock(
             config_file='nowcast.yaml')
-        m_config().__getitem__().__getitem__.return_value = {
-            'info_text': {
-                'class': 'logging.handlers.RotatingFileHandler',
-                'backupCount': 7,
-        }}
         message_broker.main()
-        handler = m_config().__getitem__().__getitem__()['info_text']
-        assert handler['class'] == 'logging.handlers.WatchedFileHandler'
-        assert 'backupCount' not in handler
         m_config().load.assert_called_once_with('nowcast.yaml')
 
+    def test_logging_config(
+        self, m_run, m_logging, m_config_logging, m_config, m_cli
+    ):
+        message_broker.main()
+        m_config_logging.assert_called_once_with(m_config())
+
     @patch('nemo_nowcast.message_broker.logger')
-    def test_logging_info(self, m_logger, m_run, m_cli, m_config, m_logging):
+    def test_logging_info(
+        self, m_logger, m_run, m_logging, m_config_logging, m_config, m_cli
+    ):
         m_cli.parser.parse_args.return_value = Mock(
             config_file='nowcast.yaml')
         m_config().load.return_value = {'logging': {}}
         message_broker.main()
-        m_logging.config.dictConfig.assert_called_once_with(
-            m_config().__getitem__())
-        assert m_logger.info.call_count == 2
+        assert m_logger.info.call_count == 3
 
-    def test_run(self, m_run, m_cli, m_config, m_logging):
+    def test_run(
+        self, m_run, m_logging, m_config_logging, m_config, m_cli
+    ):
         m_cli.parser.parse_args.return_value = Mock(
             config_file='nowcast.yaml')
         message_broker.main()
         m_run.assert_called_once_with(m_config())
+
+
+@patch('nemo_nowcast.message_broker.logging.config')
+class TestConfigureLogging:
+    """Unit tests for message_broker._configure_logging method.
+    """
+    filesystem_logging_config = {'logging': {'handlers': {'info_text': {
+        'class': 'logging.handlers.RotatingFileHandler',
+        'backupCount': 7,
+    }}}}
+    zmq_logging_config = {'logging': {
+        'publisher': {'handlers': {
+            'zmq_pub': {}}}},
+        'zmq': {
+            'host': 'localhost',
+            'ports': {'logging': {'message_broker': 4347}}}}
+
+    @pytest.mark.parametrize('config, exp_msg', [
+        (filesystem_logging_config,
+            'writing logging messages to local file system'),
+        (zmq_logging_config,
+            'publishing logging messages to tcp://*:4347'),
+    ])
+    def test_msg(self, m_logging_config, config, exp_msg):
+        msg = message_broker._configure_logging(config)
+        assert msg == exp_msg
+
+    @pytest.mark.parametrize('config', [
+        filesystem_logging_config,
+        zmq_logging_config,
+    ])
+    def test_logging_dictConfig(self, m_logging_config, config):
+        message_broker._configure_logging(config)
+        if 'publisher' in config['logging']:
+            m_logging_config.dictConfig.assert_called_once_with(
+                config['logging']['publisher'])
+        else:
+            m_logging_config.dictConfig.assert_called_once_with(
+                config['logging'])
+
+    @patch('nemo_nowcast.message_broker.logger')
+    def test_zmq_handler_root_topic(self, m_logger, m_logging_config):
+        m_handler = Mock(name='m_zmq_handler', spec=zmq.log.handlers.PUBHandler)
+        m_logger.root = Mock(handlers=[m_handler])
+        message_broker._configure_logging(self.zmq_logging_config)
+        assert m_handler.root_topic == 'message_broker'
+
+    @patch('nemo_nowcast.message_broker.logging')
+    @patch('nemo_nowcast.message_broker.logger')
+    def test_zmq_handler_formatters(
+        self, m_logger, m_logging, m_logging_config
+    ):
+        m_handler = Mock(name='m_zmq_handler', spec=zmq.log.handlers.PUBHandler)
+        m_logger.root = Mock(handlers=[m_handler])
+        message_broker._configure_logging(self.zmq_logging_config)
+        expected = {
+            m_logging.DEBUG: m_logging.Formatter("%(message)s\n"),
+            m_logging.INFO: m_logging.Formatter("%(message)s\n"),
+            m_logging.WARNING: m_logging.Formatter("%(message)s\n"),
+            m_logging.ERROR: m_logging.Formatter("%(message)s\n"),
+            m_logging.CRITICAL: m_logging.Formatter("%(message)s\n"),
+        }
+        assert m_handler.formatters == expected
+
+    def test_change_rotating_logger_handler_to_watched(self, m_logging_config):
+        message_broker._configure_logging(self.filesystem_logging_config)
+        handler = (
+            self.filesystem_logging_config['logging']['handlers']['info_text'])
+        assert handler['class'] == 'logging.handlers.WatchedFileHandler'
+        assert 'backupCount' not in handler
 
 
 @patch('nemo_nowcast.message_broker.logger')